+++ Sandswept/Items/Greens/NuclearSalvo.cs
@@ -116,6 +116,10 @@ namespace Sandswept.Items.Greens
 
             var dashRings = particles.transform.GetChild(1).GetComponent<ParticleSystemRenderer>();
             dashRings.material = newMat;
+            var dashringsPS = dashRings.GetComponent<ParticleSystem>().emission;
+            var burst = dashringsPS.GetBurst(0);
+            var burstCount = burst.count;
+            burstCount.constant = 6;
 
             var flashRed = particles.transform.GetChild(3).GetComponent<ParticleSystem>().main.startColor;
             flashRed.color = new Color32(77, 255, 0, 255);
@@ -171,7 +175,7 @@ namespace Sandswept.Items.Greens
             missileProjectileController.ghostPrefab = ghost;
 
             var missileController = missilePrefab.GetComponent<MissileController>();
-            missileController.maxSeekDistance = 10000f;
+            missileController.maxSeekDistance = 60f;
             missileController.turbulence = 0f;
             missileController.deathTimer = 30f;
             missileController.giveupTimer = 30f;
@@ -197,17 +201,16 @@ namespace Sandswept.Items.Greens
             }
 
             var inventory = body.inventory;
-            if (!inventory)
+            if (!inventory || body.GetComponent<SalvoPlayerController>())
             {
                 return;
             }
 
-            var stack = inventory.GetItemCount(ItemDef);
-            body.AddItemBehavior<SalvoPlayerController>(stack);
-            Main.ModLogger.LogError("oninventorychagned: giving itembehavior");
+            body.gameObject.AddComponent<SalvoPlayerController>();
+            // Main.ModLogger.LogError("oninventorychagned: giving itembehavior");
         }
 
-        /*
+        /* THIS IS THE PREVIOUS CODE THAT PSEUDOPULSE WROTE :SOB:
         public void GiveItem(On.RoR2.CharacterBody.orig_RecalculateStats orig, CharacterBody self)
         {
             orig(self);
@@ -275,59 +278,63 @@ namespace Sandswept.Items.Greens
         */
     }
 
-    public class SalvoPlayerController : CharacterBody.ItemBehavior
+    public class SalvoPlayerController : MonoBehaviour
     {
         public CharacterMaster master;
+        public CharacterBody body;
 
         public void Start()
         {
+            body = GetComponent<CharacterBody>();
             master = body.master;
             body.onInventoryChanged += Body_onInventoryChanged;
-        }
-
-        private void Body_onInventoryChanged()
-        {
-            var group = master.minionOwnership.group;
-            Main.ModLogger.LogError("minion ownership is " + master.minionOwnership);
-            Main.ModLogger.LogError("group is " + group);
-            Main.ModLogger.LogError("group member count is " + group.memberCount);
-            for (int i = 0; i < group.memberCount; i++)
-            {
-                var member = group.members[i];
-                var npcAlly = member.GetComponent<CharacterMaster>();
-                if (!npcAlly)
-                {
-                    Main.ModLogger.LogError("couldnt get member master");
-                    continue;
-                }
-
-                var npcBody = npcAlly.GetBody();
-                if (!npcBody)
-                {
-                    Main.ModLogger.LogError("couldnt get member body");
-                    continue;
-                }
-
-                if (npcBody && (body.bodyFlags & CharacterBody.BodyFlags.Mechanical) > CharacterBody.BodyFlags.None)
-                {
-                    Main.ModLogger.LogError("member is mechanical");
-                    var salvo = npcAlly.GetComponent<SalvoBehaviour>();
-                    if (salvo == null)
-                    {
-                        npcAlly.AddComponent<SalvoBehaviour>();
-                    }
-                    npcBody.inventory.ResetItem(NuclearSalvo.instance.ItemDef.itemIndex);
-                    npcBody.inventory.GiveItem(NuclearSalvo.instance.ItemDef, stack);
-                }
-            }
         }
 
         public void OnEnable()
         {
             if (NetworkServer.active)
             {
-                Main.ModLogger.LogError("subscribinbingign to master summon");
+                // Main.ModLogger.LogError("subscribinbingign to master summon");
                 onServerMasterSummonGlobal += MasterSummon_onServerMasterSummonGlobal;
+            }
+        }
+
+        private void Body_onInventoryChanged()
+        {
+            var members = CharacterMaster.readOnlyInstancesList.Where(member => member.minionOwnership.ownerMaster == master).ToList();
+            // Main.ModLogger.LogError("members is " + members);
+            // Main.ModLogger.LogError("member count is " + members.Count);
+            for (int i = 0; i < members.Count; i++)
+            {
+                var member = members[i];
+                var npcMaster = member.GetComponent<CharacterMaster>();
+                if (!npcMaster)
+                {
+                    // Main.ModLogger.LogError("couldnt get member master");
+                    continue;
+                }
+
+                var npcBody = npcMaster.GetBody();
+                if (!npcBody)
+                {
+                    // Main.ModLogger.LogError("couldnt get member body");
+                    continue;
+                }
+
+                if (npcBody && (npcBody.bodyFlags & CharacterBody.BodyFlags.Mechanical) > CharacterBody.BodyFlags.None)
+                {
+                    // Main.ModLogger.LogError("member is mechanical");
+                    var salvo = npcMaster.GetComponent<SalvoBehaviour>();
+                    if (salvo == null)
+                    {
+                        npcMaster.AddComponent<SalvoBehaviour>();
+                    }
+
+                    var stack = body.inventory.GetItemCount(NuclearSalvo.instance.ItemDef);
+
+                    npcBody.inventory.ResetItem(NuclearSalvo.instance.ItemDef.itemIndex);
+                    npcBody.inventory.GiveItem(NuclearSalvo.instance.ItemDef, stack);
+                }
             }
         }
 
@@ -335,35 +342,34 @@ namespace Sandswept.Items.Greens
         {
             if (!master)
             {
-                Main.ModLogger.LogError("master summon: master is not real");
                 return;
             }
 
             if (master != masterSummonReport.leaderMasterInstance)
             {
-                Main.ModLogger.LogError("master summon: leader master instance is " + masterSummonReport.leaderMasterInstance); // WHAT THE FUCK ITS NULL
-                Main.ModLogger.LogError("master summon: master doesnt equal leader master instance");
                 return;
             }
 
-            var npcAlly = masterSummonReport.summonMasterInstance;
-            if (!npcAlly)
+            var npcMaster = masterSummonReport.summonMasterInstance;
+            if (!npcMaster)
             {
-                Main.ModLogger.LogError("master summon: couldnt get summon master instance");
                 return;
             }
 
-            var npcBody = masterSummonReport.summonMasterInstance.GetBody();
-            if (npcBody && (body.bodyFlags & CharacterBody.BodyFlags.Mechanical) > CharacterBody.BodyFlags.None)
+            var npcBody = npcMaster.GetBody();
+            if (npcBody && (npcBody.bodyFlags & CharacterBody.BodyFlags.Mechanical) > CharacterBody.BodyFlags.None)
             {
-                Main.ModLogger.LogError("master summon: ally is mechanical");
-                var salvo = npcAlly.GetComponent<SalvoBehaviour>();
+                // Main.ModLogger.LogError("master summon: ally is mechanical");
+                var salvo = npcMaster.GetComponent<SalvoBehaviour>();
                 if (salvo == null)
                 {
-                    npcAlly.AddComponent<SalvoBehaviour>();
+                    npcMaster.AddComponent<SalvoBehaviour>();
                 }
-                npcBody.inventory.ResetItem(NuclearSalvo.instance.ItemDef.itemIndex);
-                npcBody.inventory.GiveItem(NuclearSalvo.instance.ItemDef, stack);
+
+                var stack = body.inventory.GetItemCount(NuclearSalvo.instance.ItemDef);
+
+                npcMaster.inventory.ResetItem(NuclearSalvo.instance.ItemDef.itemIndex);
+                npcMaster.inventory.GiveItem(NuclearSalvo.instance.ItemDef, stack);
             }
         }
 
@@ -381,8 +387,9 @@ namespace Sandswept.Items.Greens
     public class SalvoBehaviour : MonoBehaviour
     {
         public CharacterBody body;
+        public CharacterMaster master;
         public float totalMissileDelay = 5f;
-        public float enemyCheckInterval = 0.1f;
+        public float enemyCheckInterval = 0.25f;
         public float enemyCheckTimer = 0f;
         public float stopwatch = 0f;
         public bool shouldFire = false;
@@ -390,7 +397,8 @@ namespace Sandswept.Items.Greens
         public void Start()
         {
             // Main.ModLogger.LogError("salvo start");
-            body = GetComponent<CharacterBody>();
+            master = GetComponent<CharacterMaster>();
+            body = master.GetBody();
             stopwatch = totalMissileDelay;
         }
 
@@ -399,13 +407,46 @@ namespace Sandswept.Items.Greens
             stopwatch -= Time.fixedDeltaTime;
             enemyCheckTimer += Time.fixedDeltaTime;
 
-            var stack = body.inventory.GetItemCount(NuclearSalvo.instance.ItemDef);
+            var stack = master.inventory.GetItemCount(NuclearSalvo.instance.ItemDef);
+
+            if (stack <= 0)
+            {
+                Destroy(this);
+            }
 
             if (enemyCheckTimer >= enemyCheckInterval)
             {
-                for (int i = 0; i < CharacterBody.readOnlyInstancesList.Count; i++)
+                if (!body)
                 {
-                    var body = CharacterBody.readOnlyInstancesList[i];
+                    body = master.GetBody();
+                }
+
+                var sphereSearch = new SphereSearch()
+                {
+                    mask = LayerIndex.entityPrecise.mask,
+                    origin = body.corePosition,
+                    radius = 75f,
+                    queryTriggerInteraction = QueryTriggerInteraction.Ignore
+                };
+                sphereSearch.RefreshCandidates();
+                sphereSearch.FilterCandidatesByDistinctHurtBoxEntities();
+                var hurtBoxes = sphereSearch.GetHurtBoxes();
+                for (int i = 0; i < hurtBoxes.Length; i++)
+                {
+                    var hurtBox = hurtBoxes[i];
+
+                    var hc = hurtBox.healthComponent;
+                    if (!hc)
+                    {
+                        continue;
+                    }
+
+                    var body = hc.body;
+                    if (!body)
+                    {
+                        continue;
+                    }
+
                     if (!body.teamComponent)
                     {
                         continue;
@@ -418,25 +459,15 @@ namespace Sandswept.Items.Greens
 
                     shouldFire = true;
                 }
+
                 enemyCheckTimer = 0f;
             }
 
-            if (stopwatch <= 0)
+            if (stopwatch <= 0 && body && shouldFire)
             {
                 stopwatch = totalMissileDelay;
 
-                if ((body.bodyFlags & CharacterBody.BodyFlags.Mechanical) > CharacterBody.BodyFlags.None)
-                    StartCoroutine(FireMissiles(stack));
-
-                if (stack <= 0)
-                {
-                    Destroy(this);
-                }
-            }
-
-            if (stack <= 0)
-            {
-                Destroy(this);
+                StartCoroutine(FireMissiles(stack));
             }
         }