+++ Sandswept/Survivors/Ranger/Ranger.cs
@@ -16,9 +16,9 @@ namespace Sandswept.Survivors.Ranger
 
         public override string Subtitle => "Infernal Marshal";
 
-        public override string Outro => "...and so she left ready to listen to Periphery for the 43,945th time (not Periphery 3 tho it sucks it's barely replayable like she's only played it 4,874 times and got extremely fed up with it it's their most overrated album I swear)...";
+        public override string Outro => "...and so she left ready to listen to Periphery for the 43,945th time (not Periphery 3 tho it sucks it's barely replayable like she's only played it 4,874 times and got extremely fed up with it it's their most overrated album I swear)... Although she has discovered that Unprocessed may just be the better band for her overall...";
 
-        public override string Failure => "...and so she didn't leave lmao skill issue btw you should listen to these albums --- Unprocessed - In Concretion, Unprocessed - Perception, Unprocessed - Covenant, Unprocessed - And Everything In Between, Periphery - Periphery 1, Periphery - Periphery 2";
+        public override string Failure => "...and so she didn't leave lmao skill issue btw you should listen to these albums --- Unprocessed - In Concretion, Unprocessed - Perception, Unprocessed - Covenant, Unprocessed - And Everything In Between, Periphery - Periphery 1, Periphery - Periphery 2... Also you might have noticed one of Unprocessed's albums is missing - Gold - and that is because it's completely different than all their other albums, being much more pop-py and overall a sellout album.";
 
         public override void LoadAssets()
         {
@@ -91,16 +91,17 @@ namespace Sandswept.Survivors.Ranger
             ReplaceSkills(locator.utility, new SkillDef[] { Skilldefs.Sidestep.instance.skillDef });
             ReplaceSkills(locator.special, new SkillDef[] { Skilldefs.OverdriveEnter.instance.skillDef });
 
-            "SS_RANGER_BODY_LORE".Add("jaw drops\r\neyes pop out of head\r\ntongue rolls out\r\nHUMINA HUMINA HUMINA!\r\nAWOOGA AWOOGA!\r\nEE-AW EE-AW!\r\nBOIOIOING!\r\npicks up jaw\r\nfixes eyes\r\nrolls up tongue\r\nburies face in ass\r\nBLBLBLBLBL LBLBLBLBLBLBLLB\r\nWHOA MAMA");
+            "SS_RANGER_BODY_LORE".Add("After the Purge, the Hall of the Revered invested quite a lot into its own defense. For most groups, that meant shield generators, missile systems, or armies -- but the Hall's measures were more...singular.\r\n\r\nAs they had told their chosen defense when training began, the Hall was not a military organization. All they wanted was a looming threat, a force unstoppable but not flaunted or oft-used -- and unstoppable she soon became. The miracles and relics held in the trust of the Hall for eons held strength unimaginable by the known superpowers of the galaxy. \"But do not abuse them,\" she had been told. \"Our restraint and dedication to peace are the reason we were entrusted with such power, and the Hall can never betray that trust.\"\r\n\r\nAlthough her trust was firmly and eternally with the Hall and its members, the shadowed part of her brain raised doubts about this directive, echoing those words against the walls of her skull. Rather than a defense, a response to something, this felt like an attack. Sneaking aboard a UES ship was questionable to begin with; now that she was at its destination, slaughtering the denizens of the alien planet by the dozens, the doubtful questions came in floods. Out of each fatal arc of electricity and cruel burst of flame, more hesitation seeped, lodging into her mind. Had the Hall fallen? Was she receiving orders from a malicious outsider? If not, was one ancient sword, however powerful, really worth all this?");
 
             "SS_RANGER_PASSIVE_NAME".Add("Power Surge");
             "SS_RANGER_PASSIVE_DESC".Add("Hold up to " + Projectiles.DirectCurrent.maxCharge + " $rcCharge$ec. Each $rcCharge$ec increases $shbase health regeneration$se by $sh0.25 hp/s$se. $rcCharge decays over time$ec.".AutoFormat());
 
             mdl = _modelTransform.GetComponent<CharacterModel>();
 
-            var chest = _modelTransform.GetChild(0).GetChild(1).GetChild(0).GetChild(1).GetChild(0);
-            var neck = chest.GetChild(4);
-            var head = neck.GetChild(0);
+            var rig = _modelTransform.Find("Ranger Rig");
+            var chest = rig.Find("Root/Base/Stomach/Chest");
+            var neck = chest.Find("Neck");
+            var head = neck.Find("Head");
 
             var childLocator = _modelTransform.GetComponent<ChildLocator>();
             Array.Resize(ref childLocator.transformPairs, childLocator.transformPairs.Length + 3);
@@ -115,10 +116,12 @@ namespace Sandswept.Survivors.Ranger
 
             RegisterStuff();
 
+            CharacterBody.onBodyStartGlobal += CharacterBody_onBodyStartGlobal;
+
             // not sure if hgstandard has hdr emission color, but it would make the green texture pop, while still having that glow instead of being a white lightbulb with green glow
         }
 
-        private void SetupHitBox(CharacterBody body)
+        private void CharacterBody_onBodyStartGlobal(CharacterBody body)
         {
             if (body.baseNameToken != "SS_RANGER_BODY_NAME")
             {