+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # ?.?.? Patch Notes
+* Ranger
+  * Added survivor log
+  * Reworked animations, rig, IK, FK, etc
 * Items
+  * Added Crown's Diamond
+  * Added Millenium
   * Added VFX for Fractured Timepiece
   * Added item logs for Bleeding Witness and Ceremonial Jar
+  * Buffed Sun Fragment Chance 7% => 9%
+  * Nerfed Sun Fragment Stun Duration 2s => 1.5s
+  * Nerfed Nuclear Salvo Missile Range 10000m => 60m
   * Reworked Universal VIP Paws to have a chance of dropping extra items from category chests
   * Made Fractured Timepiece scale Healing and Special Cooldown Reduction with Utility Skill Base Max Stock
   * Fixed Fractured Timepiece working too well with Thermal Harpoons
+  * Fixed some things I forgot about (it's been like 3 months)
   * Optimized Nuclear Salvo
+* Elites
+  * Added Osmium (Tier 2)
+  * Improved Motivating Elite visuals
 
 # 0.8.1 Patch Notes
 * Items